+++ apps/api/src/crew.py
@@ -1,9 +1,7 @@
-import asyncio
 import logging
 import re
-from asyncio import Queue
 from typing import Any, cast
-from uuid import UUID, uuid4
+from uuid import UUID
 
 import autogen
 from autogen.cache import Cache
@@ -12,7 +10,7 @@ from src.models.session import SessionStatus
 
 from .interfaces import db
 from .models import AgentModel, CodeExecutionConfig, CrewModel, Message, Session
-from .tooling.tools import (
+from .tools import (
     generate_llm_config,
     generate_tool_from_uuid,
     get_tool_id_from_agent,
