+++ apps/api/src/tools/arxiv.py
@@ -1,0 +1,7 @@
+from langchain_community.utilities import ArxivAPIWrapper
+
+arxiv_tool = ArxivAPIWrapper()
+
+ID = ""  # TODO: Add the ID of the tool
+
+__all__ = ["arxiv_tool"]
