+++ apps/api/src/tools/test_files/cube.txt
