+++ apps/api/src/tools/test_files/square.txt
